github.com/waas_ap/requirements.txt
12factor-vault==0.1.20
aiohttp==3.8.0
asgiref==3.4.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.1.0
beautifulsoup4==4.9.1
bootstrap4==0.1.0
boto3==1.14.16
botocore==1.17.16
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.2.6
django-bootstrap4==3.0.1
django-createsuperuser==0.0.0
django-livesync==0.5
django-dbconn-retry==0.1.5
django-elasticsearch-dsl==7.1.4
django-multiform==0.1
django-vault-helpers==0.8.1
docutils==0.15.2
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
hvac==0.10.4
idna==2.10
idna-ssl==1.1.0
importlib-metadata==1.7.0
Jinja2==3.0.1
jmespath==0.10.0
python3-logstash==0.4.80
Markdown==3.2.2
MarkupSafe==2.0.1
multidict==4.7.6
pika==1.1.0
Pillow==8.3.2
portalocker==1.7.0
psycopg2-binary==2.8.5
python-consul==1.1.0
python-dateutil==2.8.1
python-logstash==0.4.6
pytz==2020.1
PyYAML==5.4.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tornado==6.0.4
typing-extensions==3.7.4.3
urllib3==1.25.9
vault-cli==1.4.0
yarl==1.5.1
zipp==3.1.0
